Make choose.dir.windows(default="<path>") actually do what it should
`?choose.dir` indicates that the `default=` argument allows the user
to indicate the folder in which the selection dialog should
open. A couple of quick checks should be enough to convince one that
it does not actually work:

```r
choose.dir("C:/Users")
choose.dir("C:/Program Files")
```

The problem is that, as documented in its next to final line, the
PowerShell script called by `choose.dir.windows()` expects an unquoted
file path:

```
```

`choose.dir.windows()`, instead, passes along a double-quoted path in
its system call to that script. That's sensible, because Windows file
paths may contain spaces, which muck up the cmd parser.

The apparently totally clean and elegant solution to this problem is
to add a simple call to the `.Trim('"')` to the bit of the PowerShell
script that processes the user-supplied `-default` argument. After
doing so, either of the above calls to `choose.dir()` work just as one
would hope.